Remove space in settings migrator csproj file
Since the build script has problems with spaces in filenames and it is quite difficult to work with batch scripts, the settings migrator will be renamed to remove spaces.

This is done in two steps in order to keep the repo nice. First the csproj file is renamed, then the folder is renamed.